Flow Controller 값 표기를 소수점 이하 2자리 형식으로 변경

+++ SapphireXR/SapphireXR_App/Models/PLCService.Definition.cs
@@ -69,7 +69,7 @@ namespace SapphireXR_App.Models
         private static Dictionary<string, ObservableManager<int>.DataIssuer>? dCurrentValueIssuers;
         private static Dictionary<string, ObservableManager<int>.DataIssuer>? dControlValueIssuers;
         private static Dictionary<string, ObservableManager<float>.DataIssuer>? dTargetValueIssuers;
-        private static Dictionary<string, ObservableManager<(int, int)>.DataIssuer>? dControlCurrentValueIssuers;
+        private static Dictionary<string, ObservableManager<(float, float)>.DataIssuer>? dControlCurrentValueIssuers;
         private static Dictionary<string, ObservableManager<float>.DataIssuer>? aMonitoringCurrentValueIssuers;
         private static ObservableManager<BitArray>.DataIssuer? baHardWiringInterlockStateIssuers;
         private static ObservableManager<BitArray>.DataIssuer? dIOStateList;
